Skyline was broken if last residue different from the previous one
The list of points used to generate the skyline intentionally has multiple entries
that share resids, which allows for the stair-step appearance. We were checking
for the last entry, in order to correctly draw the last line segment, by resid
(which could be duplicated) instead of index of the array. Now we check by index.

Hopefully fixes bug #274.